Issue #204. Refactored crypto directory
